refactor(java): align with new proto locations
